+++ examples/src/main.rs
@@ -5,38 +5,38 @@ const ITER: u32 = 1_000;
 const BENCH: bool = false;
 
 fn main() {
-    bencher("Create Server", ITER, || {
-        let _ = Server::new("127.0.0.1", 8080);
-    });
+    // bencher("Create Server", ITER, || {
+    //     let _ = Server::new("127.0.0.1", 8080);
+    // });
 
-    bencher("Add Rate limiter", ITER, || {
-        let mut server = Server::new("127.0.0.1", 8080);
-        RateLimiter::attach(&mut server, 10, 10);
-    });
+    // bencher("Add Rate limiter", ITER, || {
+    //     let mut server = Server::new("127.0.0.1", 8080);
+    //     RateLimiter::attach(&mut server, 10, 10);
+    // });
 
-    bencher("Add Logger", ITER, || {
-        let mut server = Server::new("127.0.0.1", 8080);
-        Logger::attach(
-            &mut server,
-            Logger::new(Level::Debug, Some("nose.txt"), true),
-        );
-    });
+    // bencher("Add Logger", ITER, || {
+    //     let mut server = Server::new("127.0.0.1", 8080);
+    //     Logger::attach(
+    //         &mut server,
+    //         Logger::new(Level::Debug, Some("nose.txt"), true),
+    //     );
+    // });
 
-    bencher("Add Simple Route", ITER, || {
-        let mut server = Server::new("127.0.0.1", 8080);
-        server.route(Method::GET, "/", |_| {
-            Response::new(
-                200,
-                "Hello World",
-                vec![Header::new("Content-Type", "text/plain")],
-            )
-        });
-    });
+    // bencher("Add Simple Route", ITER, || {
+    //     let mut server = Server::new("127.0.0.1", 8080);
+    //     server.route(Method::GET, "/", |_| {
+    //         Response::new(
+    //             200,
+    //             "Hello World",
+    //             vec![Header::new("Content-Type", "text/plain")],
+    //         )
+    //     });
+    // });
 
-    bencher("Add Middleware", ITER, || {
-        let mut server = Server::new("127.0.0.1", 8080);
-        server.every(Box::new(|_| None));
-    });
+    // bencher("Add Middleware", ITER, || {
+    //     let mut server = Server::new("127.0.0.1", 8080);
+    //     server.every(Box::new(|_| None));
+    // });
 
     let mut server: Server = Server::new("localhost", 1234);
 
@@ -109,6 +109,18 @@ fn main() {
             200,
             bytes.to_vec(),
             vec![Header::new("Content-Type", "image/x-icon")],
+        )
+    });
+
+    server.route(Method::GET, "/crash", |_req| {
+        panic!("CRASH!");
+    });
+
+    server.set_error_handler(|_req| {
+        Response::new(
+            500,
+            "Yikes,,, Internal Server Error",
+            vec![Header::new("Content-Type", "text/plain")],
         )
     });