+++ Changelog.md
@@ -1,3 +1,8 @@
+# 0.1.5
+- Add a route error handler
+- Add `set_error_handler` fn to set the error handler
+- Implement clone for more structs
+
 # 0.1.4
 - Allow responding with bytes not just strings
 - Add Serving Favoricon as example thing